클래스 분리
