+++ src/Game.tsx
@@ -1,44 +1,7 @@
 import { useState, useEffect } from "react";
 import { GenerateNumbers } from "./Answer";
+import { GuessInfo } from "./GuessInfo";
 
-class GuessInfo {
-  private _id?: number;
-  private _guess?: string;
-  private _ball?: number;
-  private _strike?: number;
-
-  get id(): number | undefined {
-    return this._id;
-  }
-
-  set id(value: number | undefined) {
-    this._id = value;
-  }
-
-  get guess(): string | undefined {
-    return this._guess;
-  }
-
-  set guess(value: string | undefined) {
-    this._guess = value;
-  }
-
-  get ball(): number | undefined {
-    return this._ball;
-  }
-
-  set ball(value: number | undefined) {
-    this._ball = value;
-  }
-
-  get strike(): number | undefined {
-    return this._strike;
-  }
-
-  set strike(value: number | undefined) {
-    this._strike = value;
-  }
-}
 function StartPage() {
   const [count, setCount] = useState(0);
   const [guess, setGuess] = useState("");
